Firebase user auth fix + rework Firestore design
Both app.js and auth.js have a user authentication listener to see if a user is logged in or not. It also acts as a delay to prevent window elements to be loaded from Firebase without authenticating user ID first.

Changed code for task functions so that it works with the new Firestore database design. ("users" Collection -> "uid" Doc -> "tasks" Collection)